requirements.txt
functions-framework
slack_sdk
google-cloud-secret-manager
pandas
openpyxl
dropbox
requests
